add a file for oxfordpet, fix torchvision error
