+++ python/nano/src/bigdl/nano/pytorch/vision/datasets/__init__.py
@@ -19,5 +19,6 @@ from torchvision.datasets import *
 del ImageFolder
 if not TORCHVISION_VERSION_LESS_1_12:
     del OxfordIIITPet
+    from .oxfordpet_datasets import OxfordIIITPet
 
-from .datasets import ImageFolder, SegmentationImageFolder, OxfordIIITPet
+from .datasets import ImageFolder, SegmentationImageFolder
